requirements.txt
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
dnspython==2.8.0
groq==1.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
load-dotenv==0.1.0
lxml==6.0.2
openai==2.30.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.2
sniffio==1.3.1
soupsieve==2.8.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
